graalvm-native-image-plugin-100 GraalVM 20.3.0 - extract MergeConfigFileWork interface.
